+++ streamx-flink/streamx-flink-test/src/main/scala/com/streamxhub/test/stream/DashboardApp.scala
@@ -1,8 +1,8 @@
-package com.streamxhub.stream.test
+package com.streamxhub.test.stream
 
 import com.streamxhub.flink.core.scala.{FlinkStreaming, StreamingContext}
-import org.apache.flink.streaming.api.functions.source.SourceFunction
 import org.apache.flink.api.scala._
+import org.apache.flink.streaming.api.functions.source.SourceFunction
 
 import scala.collection.JavaConverters._
 import scala.util.Random
