+++ streamx-flink/streamx-flink-test/src/main/scala/com/streamxhub/test/stream/HBaseSourceApp.scala
@@ -1,35 +1,15 @@
-/**
- * Copyright (c) 2019 The StreamX Project
- * <p>
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements. See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership. The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at
- * <p>
- * http://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied. See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package com.streamxhub.stream.test
+package com.streamxhub.test.stream
 
-import java.util
 import com.streamxhub.common.util.ConfigUtils
 import com.streamxhub.flink.core.java.wrapper.HBaseQuery
 import com.streamxhub.flink.core.scala.request.HBaseRequest
 import com.streamxhub.flink.core.scala.source.HBaseSource
 import com.streamxhub.flink.core.scala.{FlinkStreaming, StreamingContext}
 import org.apache.hadoop.hbase.CellUtil
-import org.apache.hadoop.hbase.util.Bytes
-import org.apache.flink.api.scala._
 import org.apache.hadoop.hbase.client.{Get, Scan}
+import org.apache.hadoop.hbase.util.Bytes
+
+import java.util
 
 object HBaseSourceApp extends FlinkStreaming {
