+++ streamx-flink/streamx-flink-test/src/main/scala/com/streamxhub/test/stream/KafkaSinkApp.scala
@@ -1,4 +1,4 @@
-package com.streamxhub.stream.test
+package com.streamxhub.test.stream
 
 import com.streamxhub.flink.core.scala.sink.KafkaSink
 import com.streamxhub.flink.core.scala.{FlinkStreaming, StreamingContext}
