+++ streamx-flink/streamx-flink-test/src/main/scala/com/streamxhub/test/stream/KafkaSourceApp.scala
@@ -1,8 +1,7 @@
-package com.streamxhub.stream.test
+package com.streamxhub.test.stream
 
 import com.streamxhub.flink.core.scala.source.KafkaSource
 import com.streamxhub.flink.core.scala.{FlinkStreaming, StreamingContext}
-import org.apache.flink.streaming.api.scala._
 
 object KafkaSourceApp extends FlinkStreaming {
 
@@ -26,4 +25,3 @@ object KafkaSourceApp extends FlinkStreaming {
   }
 
 }
-
