+++ streamx-flink/streamx-flink-test/src/main/scala/com/streamxhub/test/stream/MongoSourceApp.scala
@@ -1,9 +1,6 @@
-package com.streamxhub.stream.test
-
-
+package com.streamxhub.test.stream
 
 import com.mongodb.BasicDBObject
-import org.apache.flink.streaming.api.scala._
 import com.streamxhub.common.util.DateUtils
 import com.streamxhub.flink.core.scala.source.MongoSource
 import com.streamxhub.flink.core.scala.{FlinkStreaming, StreamingContext}
