+++ streamx-flink/streamx-flink-test/src/main/scala/com/streamxhub/test/stream/SideOutApp.scala
@@ -1,11 +1,10 @@
-package com.streamxhub.stream.test
+package com.streamxhub.test.stream
 
 import com.streamxhub.flink.core.scala.{FlinkStreaming, StreamingContext}
-import org.apache.flink.streaming.api.functions.source.SourceFunction
 import org.apache.flink.streaming.api.functions.ProcessFunction
-import org.apache.flink.streaming.api.scala.OutputTag
+import org.apache.flink.streaming.api.functions.source.SourceFunction
+import org.apache.flink.streaming.api.scala.{OutputTag, _}
 import org.apache.flink.util.Collector
-import org.apache.flink.streaming.api.scala._
 
 import scala.util.Random
