+++ streamx-flink/streamx-flink-test/src/main/scala/com/streamxhub/test/stream/StreamingTestApp.scala
@@ -1,8 +1,7 @@
-package com.streamxhub.stream.test
+package com.streamxhub.test.stream
 
 import com.streamxhub.flink.core.scala.{FlinkStreaming, StreamingContext}
 import org.apache.flink.api.common.functions.ReduceFunction
-import org.apache.flink.api.scala._
 
 import scala.util.Try
